# refactor: reduced closure duplication using thread id