new `getUniqueThreadClosure` function:
```java
public Clo1%Obj%Obj getUniqueThreadClosure() {
        if (Thread.currentThread().getId() == this.threadId) {
            return this;
        } else {
            Clo%main%121576 var10000 = new Clo%main%121576();
            var10000.threadId = Thread.currentThread().getId();
            return var10000;
        }
    }
```