@JonathanStarup 2021-11-29
# PR
====================== Flix Benchmark ======================
. | . 
:--- | ---:
BenchmarkList.benchmark01 | 23.298 ns.
BenchmarkList.benchmark02 | 34.108 ns.
BenchmarkList.benchmark03 | 26.147 ns.
BenchmarkList.benchmark04 | 35.671 ns.
BenchmarkList.benchmark05 | 44.052 ns.
BenchmarkList.benchmark06 | 85.564 ns.
BenchmarkList.benchmark07 | 53.655 ns.
BenchmarkList.benchmark08 | 27.273 ns.
BenchmarkList.benchmark09 | 39.577 ns.
BenchmarkList.benchmark10 | 51.200 ns.
BenchmarkList.benchmark11 | 78.326 ns.
BenchmarkList.benchmark12 | 103.961 ns.
BenchmarkList.benchmark13 | 32.500 ns.
BenchmarkList.benchmark14 | 28.761 ns.
BenchmarkList.benchmark15 | 68.112 ns.
BenchmarkList.benchmark16 | 134.376 ns.
BenchmarkList.benchmark17 | 30.110 ns.
BenchmarkList.benchmark18 | 53.612 ns.

Finished 18 benchmarks with 25000 iterations each.