@JonathanStarup 2021-11-29
@magnus-madsen looks good for performance even in this version