@magnus-madsen 2021-11-29
Yes, so seems like worth doing.