@magnus-madsen 2021-11-29
@JonathanStarup Did you test it on parMap for RBTree? :)