@JonathanStarup 2021-11-30
> @JonathanStarup Did you test it on parMap for RBTree? :)

[This](https://github.com/flix/flix/issues/2336#issuecomment-980541504) works fine if that's what you mean