@magnus-madsen 2021-11-30
Could you also post the impact on the Datalog engine? Thanks!