@JonathanStarup 2021-11-30
> Could you also post the impact on the Datalog engine? Thanks!

Benchmark | Iterations(PR) | Time (avg)(PR) | Iterations(Master) | Time (avg)(Master)
:--- | ---: | ---: | ---: | ---:
Project (Int, Int) into A (n = 16) | 114103 | 15 us/op | 141442 | 13 us/op
Project (Int, Int) into A (n = 256) | 35483 | 416 us/op | 37832 | 381 us/op
Project (Int, Int) into A (n = 4096) | 1713 | 11319 us/op | 1606 | 9321 us/op
Project (Int, Int) into A (n = 65536) | 76 | 213013 us/op | 74 | 207281 us/op
Project (Int, Int) into A (n = 1048576) | 3 | 4472140 us/op | 3 | 4348798 us/op
Select (Int, Int) from A (n = 16) | 2571 | 277 us/op | 3030 | 201 us/op
Select (Int, Int) from A (n = 256) | 7260 | 140 us/op | 9320 | 121 us/op
Select (Int, Int) from A (n = 4096) | 2126 | 424 us/op | 1887 | 388 us/op
Select (Int, Int) from A (n = 65536) | 352 | 6573 us/op | 333 | 6498 us/op
Select (Int, Int) from A (n = 1048576) | 86 | 208482 us/op | 107 | 187040 us/op
Solve (Closure) (n = 64) | 2076 | 601 us/op | 2094 | 383 us/op
Solve (Closure) (n = 256) | 1027 | 751 us/op | 971 | 615 us/op
Solve (Closure) (n = 1024) | 428 | 2740 us/op | 311 | 2397 us/op
Solve (Closure) (n = 4096) | 141 | 11771 us/op | 144 | 11001 us/op
Solve (Closure) (n = 16384) | 59 | 53764 us/op | 60 | 52300 us/op
Solve (Closure) (n = 65536) | 32 | 259256 us/op | 27 | 255597 us/op
Solve (Closure) (n = 262144) | 9 | 1302932 us/op | 11 | 1313500 us/op
Solve (Cross Product 2) (n = 256) | 3644 | 560 us/op | 6873 | 503 us/op
Solve (Cross Product 2) (n = 1024) | 2047 | 2486 us/op | 3161 | 2356 us/op
Solve (Cross Product 2) (n = 4096) | 653 | 11960 us/op | 239 | 11508 us/op
Solve (Cross Product 2) (n = 16384) | 239 | 60201 us/op | 232 | 54675 us/op
Solve (Cross Product 2) (n = 65536) | 57 | 260331 us/op | 49 | 249499 us/op
Solve (Cross Product 2) (n = 262144) | 12 | 1189148 us/op | 12 | 1138940 us/op
Solve (Cross Product 2) (n = 1048576) | 2 | 5500383 us/op | 2 | 5175403 us/op
Solve (Cross Product 3) (n = 4096) | 1137 | 11905 us/op | 1267 | 10862 us/op
Solve (Cross Product 3) (n = 32768) | 113 | 127763 us/op | 124 | 122686 us/op
Solve (Cross Product 3) (n = 262144) | 10 | 1240251 us/op | 11 | 1221653 us/op
Solve (Intersect) (n = 16) | 8022 | 64 us/op | 6681 | 76 us/op
Solve (Intersect) (n = 256) | 6557 | 606 us/op | 6844 | 646 us/op
Solve (Intersect) (n = 4096) | 960 | 13603 us/op | 992 | 14718 us/op
Solve (Intersect) (n = 65536) | 53 | 268487 us/op | 55 | 260609 us/op
Solve (Intersect) (n = 1048576) | 2 | 5273125 us/op | 2 | 5198168 us/op
Compile Rules (n = 16) | 492 | 1149 us/op | 610 | 1116 us/op
Compile Rules (n = 256) | 52 | 20353 us/op | 49 | 18791 us/op
Compile Rules (n = 4096) | 9 | 335207 us/op | 9 | 310581 us/op