@magnus-madsen 2021-11-30
What's the conclusion? That it's a tiny bit slower?