@JonathanStarup 2021-11-30
> What's the conclusion? That it's a tiny bit slower?

yeah, but also faster on list. takes more benchmarking if we want the exact effect I think.