@magnus-madsen 2021-11-30
I'd like you to explore two things:

- If it helps to cache the current thread id.
- What the generated machine code looks like. 


The latter should be fun to explore. The starting point is a tiny program and the flags `-XX:+PrintCompilation`, `-XX:-TieredCompilation`, `-XX:+PrintCompilation`, and `-XX:+UnlockDiagnosticVMOptions -XX:+PrintAssembly`.