@magnus-madsen 2021-12-13
Wait until `List` functions have been refactor to use a local `loop` and then re-benchmark.