@magnus-madsen 2021-12-26
List has now been refactored, if you want to redo the experiments.