@magnus-madsen 2022-07-02
I am closing this. I think a better way is to ensure our local letrecs are rewritten to loops-- when possible.